static data inserted

+++ src/main/java/usm/api/doctoral_registration/repository/SciencesBranchRepository.java
@@ -2,8 +2,8 @@ package usm.api.doctoral_registration.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import usm.api.doctoral_registration.model.school.SciencesBranch;
+import usm.api.doctoral_registration.model.sciences.SciencesBranch;
 
 @Repository
-public interface SciencesBranchRepository extends JpaRepository<SciencesBranch, Long> {
+public interface SciencesBranchRepository extends JpaRepository<SciencesBranch, Float> {
 }
