+++ src/main/java/usm/api/doctoral_registration/repository/SciencesDomainRepository.java
@@ -2,8 +2,8 @@ package usm.api.doctoral_registration.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import usm.api.doctoral_registration.model.school.SciencesDomain;
+import usm.api.doctoral_registration.model.sciences.SciencesDomain;
 
 @Repository
-public interface SciencesDomainRepository extends JpaRepository<SciencesDomain, Long> {
+public interface SciencesDomainRepository extends JpaRepository<SciencesDomain, Float> {
 }
