+++ src/main/java/usm/api/doctoral_registration/repository/SpecialityRepository.java
@@ -2,8 +2,8 @@ package usm.api.doctoral_registration.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import usm.api.doctoral_registration.model.school.Speciality;
+import usm.api.doctoral_registration.model.sciences.Speciality;
 
 @Repository
-public interface SpecialityRepository extends JpaRepository<Speciality, Long> {
+public interface SpecialityRepository extends JpaRepository<Speciality, Float> {
 }
