+++ src/main/resources/db/migration/V01__create_table_speciality.sql
@@ -2,9 +2,8 @@ CREATE SEQUENCE IF NOT EXISTS hibernate_sequence START WITH 1 INCREMENT BY 1;
 
 CREATE TABLE speciality
 (
-    id                  BIGINT NOT NULL,
+    id                  FLOAT NOT NULL,
     name                VARCHAR(255),
-    sciences_profile_id BIGINT,
+    sciences_profile_id FLOAT,
     CONSTRAINT pk_speciality PRIMARY KEY (id)
 );
-
