+++ src/main/resources/db/migration/V02__create_table_sciences_profile_add_constraint.sql
@@ -2,9 +2,9 @@ CREATE SEQUENCE IF NOT EXISTS hibernate_sequence START WITH 1 INCREMENT BY 1;
 
 CREATE TABLE sciences_profile
 (
-    id                 BIGINT NOT NULL,
+    id                 FLOAT NOT NULL,
     name               VARCHAR(255),
-    sciences_branch_id BIGINT,
+    sciences_branch_id FLOAT,
     CONSTRAINT pk_sciences_profile PRIMARY KEY (id)
 );
